Remove MainMessage fieldset
