+++ pkg/task/inspection/googlecloudcommon/contract/fieldset_test.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/GoogleCloudPlatform/khi/pkg/common/structured"
 	pb "github.com/GoogleCloudPlatform/khi/pkg/generated/khifile/v6"
+	"github.com/GoogleCloudPlatform/khi/pkg/model/log"
 	commonlogk8saudit_contract "github.com/GoogleCloudPlatform/khi/pkg/task/inspection/commonlogk8saudit/contract"
 	inspectioncore_contract "github.com/GoogleCloudPlatform/khi/pkg/task/inspection/inspectioncore/contract"
 )
@@ -166,4 +167,101 @@ func TestGCPDefaultSeverityFieldSetReader(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestGCPMainMessageFieldSet(t *testing.T) {
+	testCase := []struct {
+		Name                string
+		ExpectedMainMessage string
+		InputYAML           string
+	}{
+		{
+			Name:                "from textPayload field",
+			ExpectedMainMessage: "foo",
+			InputYAML:           `textPayload: foo`,
+		},
+		{
+			Name:                "from jsonPayload.message field",
+			ExpectedMainMessage: "bar",
+			InputYAML: `jsonPayload:
+  message: bar`,
+		},
+		{
+			Name:                "from jsonPayload.MESSAGE field",
+			ExpectedMainMessage: "bar",
+			InputYAML: `jsonPayload:
+  MESSAGE: bar`,
+		},
+		{
+			Name:                "from jsonPayload.msg field",
+			ExpectedMainMessage: "bar",
+			InputYAML: `jsonPayload:
+  msg: bar`,
+		},
+		{
+			Name:                "from jsonPayload.log field",
+			ExpectedMainMessage: "bar",
+			InputYAML: `jsonPayload:
+  log: bar`,
+		},
+		{
+			Name:                "from the whole jsonPayload field",
+			ExpectedMainMessage: `{"foo":"bar"}`,
+			InputYAML: `jsonPayload:
+  foo: bar`,
+		},
+		{
+			Name:                "from the whole labels field",
+			ExpectedMainMessage: `{"foo":"bar"}`,
+			InputYAML: `labels:
+  foo: bar`,
+		},
+		{
+			Name:                "ignore when the message is protoPayload even labels are provided",
+			ExpectedMainMessage: "",
+			InputYAML: `labels:
+  foo: bar
+protoPayload:
+  qux: quux`,
+		},
+		{
+			Name:                "empty if no proper field is given",
+			ExpectedMainMessage: "",
+			InputYAML:           `foo: bar`,
+		},
+		{
+			Name:                "prioritize textPayload rather than jsonPayload.msg or labels",
+			ExpectedMainMessage: "bar",
+			InputYAML: `jsonPayload:
+  msg: foo
+textPayload: bar
+labels:
+  qux: quux`,
+		},
+		{
+			Name:                "prioritize jsonPayload.msg over labels",
+			ExpectedMainMessage: "foo",
+			InputYAML: `jsonPayload:
+  msg: foo
+labels:
+  qux: quux`,
+		},
+	}
+	for _, tc := range testCase {
+		t.Run(tc.Name, func(t *testing.T) {
+			l, err := log.NewLogFromYAMLString(tc.InputYAML)
+			if err != nil {
+				t.Fatalf("failed to parse log from yaml: %v", err)
+			}
+			l.SetFieldSetReader(&GCPMainMessageFieldSetReader{})
+			gcpMainMessageField, err := log.GetFieldSet(l, &GCPMainMessageFieldSet{})
+			if err != nil {
+				t.Fatalf("failed to extract gcp main message field: %v", err)
+			}
+			if gcpMainMessageField.MainMessage != tc.ExpectedMainMessage {
+				t.Errorf("expected main message: %v, got: %v", tc.ExpectedMainMessage, gcpMainMessageField.MainMessage)
+			}
+		})
+	}
+
 }
